move inversion config to runners package
